# src/swagger.json
{
  "openapi":"3.0.0",
  "info": {
    "title": "Api Controle de usuário",
    "description": "Essa API tem como objetivo realizar controle de usuários através de token JWT",
    "termsOfService": "http://localhost:3000/terms-of-service",
    "contact": {
      "email": "michelmpin@gmail.com"
    },
    "version": "1.0.0"
  },
  "servers":[
    {
      "url": "http://localhost:3000",
      "description": "API de Teste"
    }
  ],
  "paths": {
    "/login": {
      "post": {
        "summary": "Efetuar login",
        "description": "Essa rota é responsável por efetuar o login de usuário retornando um token no formato JWT com um id para um refresh token",
        "tags": ["Login"],
        "securiy": [{"bearerAuth":[]}],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Login"
              },
              "examples": {
                "login": {
                  "value": {
                    "username": "mick",
                    "password": "1234"
                  }
                }
              }
            }
          }
        },
        "responses":{
          "401": {
            "description": "User or password incorrect "
          },
          "200": {
            "description": "Ok",
            "content":{
              "application/json": {
                "schema": {
                  "type":"object",
                  "$ref": "#/components/schemas/Token"
                },
                "examples": {
                  "login": {
                    "value": {
                      "token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpYXQiOjE2NDkyNjk5MjgsImV4cCI6MTY0OTI2OTk0OCwic3ViIjoiMWRmNzk0NDQtNGYzNi00N2YzLWI3MjUtYzVjNzc4ZDFjNDY2In0.D6xZM04zcEumxgowBrwNaCaxX1PfDzCsUKgqWhgsl3M",
                      "refreshToken": {
                        "id": "1f2d195c-4198-4d88-b217-a324211aa0a3",
                        "expiresIn": 1649269948,
                        "userId": "1df79444-4f36-47f3-b725-c5c778d1c466"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/token": {
      "get": {
        "summary": "Verifica token",
        "dcescription": "Essa rota é responsável por verificar se o token enviado é válido",
        "tags": ["Validation"],
        "security": [{"bearerAuth":[]}],
        "responses":{
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type":"string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description":"Unauthorized",
            "content":{
              "application/json": {
                "schema":{
                  "type":"object",
                  "properties":{
                    "message":{
                      "type":"string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Login": {
        "type": "object",
        "properties": {
          "username":{
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "Token": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          },
          "refreshToken": {
            "type":"object",
            "properties":{
              "id": {
                "type":"string"
              },
              "expiresIn": {
                "type":"number"
              },
              "user-id": {
                "type":"string"
              }
            }
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}